tests: Merge the hashmap test files
