+++ tests/coq/hashmap/Hashmap_FunsExternal_Template.v
@@ -1,0 +1,23 @@
+(** THIS FILE WAS AUTOMATICALLY GENERATED BY AENEAS *)
+(** [hashmap]: external functions.
+-- This is a template file: rename it to "FunsExternal.lean" and fill the holes. *)
+Require Import Primitives.
+Import Primitives.
+Require Import Coq.ZArith.ZArith.
+Require Import List.
+Import ListNotations.
+Local Open Scope Primitives_scope.
+Require Import Hashmap_Types.
+Include Hashmap_Types.
+Module Hashmap_FunsExternal_Template.
+
+(** [hashmap::hashmap_utils::deserialize]:
+    Source: 'tests/src/hashmap.rs', lines 330:4-330:47 *)
+Axiom hashmap_utils_deserialize : state -> result (state * (HashMap_t u64)).
+
+(** [hashmap::hashmap_utils::serialize]:
+    Source: 'tests/src/hashmap.rs', lines 325:4-325:46 *)
+Axiom hashmap_utils_serialize : HashMap_t u64 -> state -> result (state * unit)
+.
+
+End Hashmap_FunsExternal_Template.
