+++ tests/coq/hashmap/Hashmap_TypesExternal.v
@@ -1,14 +1,13 @@
-(** [hashmap_main]: external types.
--- This is a template file: rename it to "TypesExternal.lean" and fill the holes. *)
+(** [hashmap]: external types. *)
 Require Import Primitives.
 Import Primitives.
 Require Import Coq.ZArith.ZArith.
 Require Import List.
 Import ListNotations.
 Local Open Scope Primitives_scope.
-Module HashmapMain_TypesExternal.
+Module Hashmap_TypesExternal.
 
 (** The state type used in the state-error monad *)
 Axiom state : Type.
 
-End HashmapMain_TypesExternal.
+End Hashmap_TypesExternal.
