+++ tests/coq/hashmap/Hashmap_TypesExternal_Template.v
@@ -1,5 +1,5 @@
 (** THIS FILE WAS AUTOMATICALLY GENERATED BY AENEAS *)
-(** [hashmap_main]: external types.
+(** [hashmap]: external types.
 -- This is a template file: rename it to "TypesExternal.lean" and fill the holes. *)
 Require Import Primitives.
 Import Primitives.
@@ -7,9 +7,9 @@ Require Import Coq.ZArith.ZArith.
 Require Import List.
 Import ListNotations.
 Local Open Scope Primitives_scope.
-Module HashmapMain_TypesExternal_Template.
+Module Hashmap_TypesExternal_Template.
 
 (** The state type used in the state-error monad *)
 Axiom state : Type.
 
-End HashmapMain_TypesExternal_Template.
+End Hashmap_TypesExternal_Template.
