--- tests/coq/hashmap_main/HashmapMain_FunsExternal_Template.v
@@ -1,26 +1,0 @@
-(** THIS FILE WAS AUTOMATICALLY GENERATED BY AENEAS *)
-(** [hashmap_main]: external functions.
--- This is a template file: rename it to "FunsExternal.lean" and fill the holes. *)
-Require Import Primitives.
-Import Primitives.
-Require Import Coq.ZArith.ZArith.
-Require Import List.
-Import ListNotations.
-Local Open Scope Primitives_scope.
-Require Import HashmapMain_Types.
-Include HashmapMain_Types.
-Module HashmapMain_FunsExternal_Template.
-
-(** [hashmap_main::hashmap_utils::deserialize]:
-    Source: 'tests/src/hashmap_utils.rs', lines 11:0-11:43 *)
-Axiom hashmap_utils_deserialize
-  : state -> result (state * (hashmap_HashMap_t u64))
-.
-
-(** [hashmap_main::hashmap_utils::serialize]:
-    Source: 'tests/src/hashmap_utils.rs', lines 6:0-6:42 *)
-Axiom hashmap_utils_serialize
-  : hashmap_HashMap_t u64 -> state -> result (state * unit)
-.
-
-End HashmapMain_FunsExternal_Template.
