--- tests/coq/hashmap_main/_CoqProject
@@ -1,12 +1,0 @@
-# This file was automatically generated - see ../Makefile
--R . Lib
--arg -w
--arg all
-
-HashmapMain_Types.v
-HashmapMain_FunsExternal_Template.v
-Primitives.v
-HashmapMain_Funs.v
-HashmapMain_TypesExternal.v
-HashmapMain_FunsExternal.v
-HashmapMain_TypesExternal_Template.v
