+++ tests/fstar/hashmap/Hashmap.FunsExternal.fsti
@@ -1,0 +1,16 @@
+(** THIS FILE WAS AUTOMATICALLY GENERATED BY AENEAS *)
+(** [hashmap]: external function declarations *)
+module Hashmap.FunsExternal
+open Primitives
+include Hashmap.Types
+
+#set-options "--z3rlimit 50 --fuel 1 --ifuel 1"
+
+(** [hashmap::hashmap_utils::deserialize]:
+    Source: 'tests/src/hashmap.rs', lines 330:4-330:47 *)
+val hashmap_utils_deserialize : state -> result (state & (hashMap_t u64))
+
+(** [hashmap::hashmap_utils::serialize]:
+    Source: 'tests/src/hashmap.rs', lines 325:4-325:46 *)
+val hashmap_utils_serialize : hashMap_t u64 -> state -> result (state & unit)
+
