+++ tests/fstar/hashmap/Hashmap.TypesExternal.fsti
@@ -1,6 +1,6 @@
 (** THIS FILE WAS AUTOMATICALLY GENERATED BY AENEAS *)
-(** [hashmap_main]: external type declarations *)
-module HashmapMain.TypesExternal
+(** [hashmap]: external type declarations *)
+module Hashmap.TypesExternal
 open Primitives
 
 #set-options "--z3rlimit 50 --fuel 1 --ifuel 1"
