--- tests/fstar/hashmap_main/HashmapMain.FunsExternal.fsti
@@ -1,18 +1,0 @@
-(** THIS FILE WAS AUTOMATICALLY GENERATED BY AENEAS *)
-(** [hashmap_main]: external function declarations *)
-module HashmapMain.FunsExternal
-open Primitives
-include HashmapMain.Types
-
-#set-options "--z3rlimit 50 --fuel 1 --ifuel 1"
-
-(** [hashmap_main::hashmap_utils::deserialize]:
-    Source: 'tests/src/hashmap_utils.rs', lines 11:0-11:43 *)
-val hashmap_utils_deserialize
-  : state -> result (state & (hashmap_HashMap_t u64))
-
-(** [hashmap_main::hashmap_utils::serialize]:
-    Source: 'tests/src/hashmap_utils.rs', lines 6:0-6:42 *)
-val hashmap_utils_serialize
-  : hashmap_HashMap_t u64 -> state -> result (state & unit)
-
