+++ tests/lean/Hashmap/FunsExternal_Template.lean
@@ -1,0 +1,17 @@
+-- THIS FILE WAS AUTOMATICALLY GENERATED BY AENEAS
+-- [hashmap]: external functions.
+-- This is a template file: rename it to "FunsExternal.lean" and fill the holes.
+import Base
+import Hashmap.Types
+open Primitives
+open hashmap
+
+/- [hashmap::hashmap_utils::deserialize]:
+   Source: 'tests/src/hashmap.rs', lines 330:4-330:47 -/
+axiom hashmap_utils.deserialize : State → Result (State × (HashMap U64))
+
+/- [hashmap::hashmap_utils::serialize]:
+   Source: 'tests/src/hashmap.rs', lines 325:4-325:46 -/
+axiom hashmap_utils.serialize
+  : HashMap U64 → State → Result (State × Unit)
+
