+++ tests/lean/Hashmap/Types.lean
@@ -1,18 +1,19 @@
 -- THIS FILE WAS AUTOMATICALLY GENERATED BY AENEAS
 -- [hashmap]: type definitions
 import Base
+import Hashmap.TypesExternal
 open Primitives
 
 namespace hashmap
 
 /- [hashmap::List]
-   Source: 'tests/src/hashmap.rs', lines 27:0-27:16 -/
+   Source: 'tests/src/hashmap.rs', lines 29:0-29:16 -/
 inductive List (T : Type) :=
 | Cons : Usize → T → List T → List T
 | Nil : List T
 
 /- [hashmap::HashMap]
-   Source: 'tests/src/hashmap.rs', lines 43:0-43:21 -/
+   Source: 'tests/src/hashmap.rs', lines 45:0-45:21 -/
 structure HashMap (T : Type) where
   num_entries : Usize
   max_load_factor : (Usize × Usize)
