+++ tests/lean/Hashmap/TypesExternal_Template.lean
@@ -1,5 +1,5 @@
 -- THIS FILE WAS AUTOMATICALLY GENERATED BY AENEAS
--- [hashmap_main]: external types.
+-- [hashmap]: external types.
 -- This is a template file: rename it to "TypesExternal.lean" and fill the holes.
 import Base
 open Primitives
