--- tests/lean/HashmapMain.lean
@@ -1,1 +1,0 @@
-import HashmapMain.Funs
