--- tests/lean/HashmapMain/FunsExternal_Template.lean
@@ -1,18 +1,0 @@
--- THIS FILE WAS AUTOMATICALLY GENERATED BY AENEAS
--- [hashmap_main]: external functions.
--- This is a template file: rename it to "FunsExternal.lean" and fill the holes.
-import Base
-import HashmapMain.Types
-open Primitives
-open hashmap_main
-
-/- [hashmap_main::hashmap_utils::deserialize]:
-   Source: 'tests/src/hashmap_utils.rs', lines 11:0-11:43 -/
-axiom hashmap_utils.deserialize
-  : State → Result (State × (hashmap.HashMap U64))
-
-/- [hashmap_main::hashmap_utils::serialize]:
-   Source: 'tests/src/hashmap_utils.rs', lines 6:0-6:42 -/
-axiom hashmap_utils.serialize
-  : hashmap.HashMap U64 → State → Result (State × Unit)
-
