--- tests/lean/HashmapMain/Types.lean
@@ -1,23 +1,0 @@
--- THIS FILE WAS AUTOMATICALLY GENERATED BY AENEAS
--- [hashmap_main]: type definitions
-import Base
-import HashmapMain.TypesExternal
-open Primitives
-
-namespace hashmap_main
-
-/- [hashmap_main::hashmap::List]
-   Source: 'tests/src/hashmap.rs', lines 27:0-27:16 -/
-inductive hashmap.List (T : Type) :=
-| Cons : Usize → T → hashmap.List T → hashmap.List T
-| Nil : hashmap.List T
-
-/- [hashmap_main::hashmap::HashMap]
-   Source: 'tests/src/hashmap.rs', lines 43:0-43:21 -/
-structure hashmap.HashMap (T : Type) where
-  num_entries : Usize
-  max_load_factor : (Usize × Usize)
-  max_load : Usize
-  slots : alloc.vec.Vec (hashmap.List T)
-
-end hashmap_main
