+++ tests/lean/lakefile.lean
@@ -13,7 +13,6 @@ package «tests» {}
 @[default_target] lean_lib Constants
 @[default_target] lean_lib External
 @[default_target] lean_lib Hashmap
-@[default_target] lean_lib HashmapMain
 @[default_target] lean_lib Loops
 @[default_target] lean_lib NoNestedBorrows
 @[default_target] lean_lib Paper
